requirements/prod.txt
# requirements/prod.txt
-r common.txt
django-compressor==1.1.2
django-htmlmin==0.5.1
gunicorn==0.14.2
psycopg2==2.4.5
raven==1.3.5
dj-database-url